fix: /chat/stream always uses web grounding, drop BigQuery routing
Financial-metric routing in AgentV2 silently dropped out-of-range years
during query parsing (e.g. "2025"), so BigQuery ran with no year filter,
returned all-years data, and the router never fell through to Gemini's
web-search grounding as intended -- the chatbot reported "no data" for
years it should have searched the web for. Rather than keep patching
that path, /chat/stream now always answers via Gemini 2.5 Pro + Google
Search grounding; only the cheap Flash REFUSE safety pre-check remains
in the router. financial_tool.py had no other callers once that branch
was removed, so it (and its tests) were deleted.

FinancialDataManager.parse_user_query (shared with /fast_chat_v2) still
gets the year-passthrough fix, since /fast_chat_v2 has the same bug.
